requirements.txt
streamlit>=1.28.0
anthropic>=0.18.0
openai>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
